requirements.txt
PyOpenGL==3.1.7
numpy==1.24.3
Pillow==10.1.0
pygame==2.5.2